renamed models to classes
